shuffled around some mod stuff to make everything fit properly.
