+++ decompile/General/MAIN/MainFrame_RenderFrame.c
@@ -1832,9 +1832,12 @@ void RenderSubmit(struct GameTracker* gGT)
 
 	void* ot = &gGT->pushBuffer[0].ptrOT[0x3ff];
 
-#if 0 && defined(USE_ONLINE)
-	void OnlineMirrorMode(u_long* ot);
-	OnlineMirrorMode(ot);
+#if defined(USE_ONLINE) && 0
+	if ((octr->onlineGameModifiers & MODIFIER_MIRROR))
+	{
+		void OnlineMirrorMode(u_long* ot);
+		OnlineMirrorMode(ot);
+	}
 #endif
 
 	DrawOTag(ot);
